Extracted client type images, better padding for elements
